functions/process/requirements.txt
psycopg2-binary==2.9.9
yt-dlp
ffmpeg
boto3